+++ frontend/src/functionality/CreateSchedule/utils/Api/generateSchedule.js
@@ -1,4 +1,4 @@
-import { api } from '../../api.js';
+import { api } from '../../../../api.js';
 
 /**
  * Generates a schedule
