+++ frontend/src/functionality/CreateSchedule/utils/Hooks/useGenerateSchedule.js
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import generateSchedule from '../../../../utils/Api/generateSchedule.js';
+import generateSchedule from '../Api/generateSchedule.js';
 
 
 /**
